+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Show when messages are from a thread
+- Add command `/rocketchat reply <tmid> <message>` to reply inside a thread
+
 ## [0.4.0] - 2022-03-12
 
 ### Added